c: revise day 2 part 1
Pass the array of policies and the length, then iterate by incrementing
the pointer, in the idiomatic way.

Also, the three line `sscanf` solution to parsing the input data is
so, so much easier than regex. And it's fine as long as all of the
input data 100% matches a single format.